Move functions to their right locations
